Refactor code to reduce redundancy
