+++ src/js/api/posts/getPosts.mjs
@@ -1,27 +1,12 @@
-import { API_KEY } from "../constants.mjs";
-import { load } from "../../storage/localStorage.mjs";
+import { fetchData } from "../GET/fetchData.mjs";
 import { renderPosts } from "./renderPosts.mjs";
-
-const method = "GET";
 
 export async function getPosts(url) {
   try {
-    const token = load("token");
-
-    const getData = {
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${token}`,
-        "X-Noroff-API-Key": API_KEY,
-      },
-      method,
-    };
-    const response = await fetch(url, getData);
-    const result = await response.json();
+    const result = await fetchData(url);
     console.log("JSON/RESULT: ", result);
-
     renderPosts(result);
   } catch (error) {
-    console.log("ERROR:", error);
+    console.log("ERROR FROM GET POSTS: ", error);
   }
 }
